fix(FIX-020): consolidate duplicate ErrorBoundary into single root-level component
Delete components/ErrorBoundary.tsx (ISSUE-MICRO-105), keep src/ErrorBoundary.tsx
(GO-LIVE-172) which has dev error details + Go to Home button.
Update App.tsx and test imports to use root-level ErrorBoundary.
